Broadcast Bot
The bot can be useful for such development groups/ movie groups or any other groups that they can broadcast official messages and media to the followers who started the bot. By including this repo code in public bots can beneficial for the owners to  push  the ads or messages to the bot users.